backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pydantic>=2.9.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
asyncpg>=0.29.0
sqlalchemy>=2.0.0
alembic>=1.13.0
httpx>=0.27.0
redis>=5.0.0
python-dotenv>=1.0.0
python-multipart>=0.0.9
psycopg2-binary>=2.9.0
email-validator==2.1.0.post1
pydantic-settings>=2.0.0
pandas>=2.0.0
requests>=2.31.0
torch>=2.0.0
transformers>=4.30.0
fugashi>=1.3.0
ipadic>=1.0.0
